Fixed a compilation issue where the typedef for GLsizei does not match the OpenGL header type
Fix: 978

+++ include/CoreVideo/CVOpenGLESTexture.h
@@ -27,7 +27,7 @@ typedef unsigned int GLuint;
 typedef signed int GLint;
 typedef char GLchar;
 typedef float GLfloat;
-typedef size_t GLsizei;
+typedef int GLsizei;
 
 typedef CVImageBufferRef CVOpenGLESTextureRef;
